build: Convert storage.js to ts.
